git-svn-id: http://sonetto.googlecode.com/svn@223 0bea0c47-0c44-0410-8422-6f9b8df7ced0

+++ trunk/libsonetto/include/SonettoJoystick.h
@@ -77,14 +77,14 @@ namespace Sonetto
             RWA_2
         };
 
-        Joystick(uint16 id) : mID(id),mJoy(NULL) {}
+        Joystick(uint32 id) : mID(id),mJoy(NULL) {}
 
         virtual ~Joystick();
 
         // Winmm hack
         bool isPlugged();
 
-        inline uint16 getID() const { return mID; }
+        inline uint32 getID() const { return mID; }
 
         inline bool isEnabled() const { return (mJoy != NULL); }
 
@@ -95,7 +95,7 @@ namespace Sonetto
         Ogre::Vector2 getRawAnalogState(RawAnalog analog);
 
     protected:
-        uint16 mID;
+        uint32 mID;
 
         SDL_Joystick *mJoy;
     };
